Break deck tests out into a module for reuse for shufflers
