+++ test/cards.js
@@ -1,21 +1,9 @@
 var expect = require('chai').expect;
 var Cards = require('../lib/cards');
+var deckTest = require('./deck');
 
 describe('makeDeck', function(){
-	it('returns 52 cards', function(){
-		var deck = Cards.makeDeck();
-		expect(deck.length).to.equal(52);
-	});
-
-	it('returns no repeated cards', function(){
-		var deck = Cards.makeDeck();
-		var cards = {};
-		for (var i = 0; i< deck.length; i++) {
-			var cardName = deck[i].getName();
-			expect(cards[cardName]).a('undefined');
-			cards[cardName] = "already exists";;
-		}
-	});
+		deckTest.runDeckTests(Cards.makeDeck());
 });
 
 describe('getName and ranks', function() {
